+++ include/xtal--math/process/pade/unity.hh
@@ -80,7 +80,7 @@ struct unity<M_ism> : unity<>
 		}
 
 		template <int N_lim=-1>
-		XTAL_DEF_(long,static)
+		XTAL_DEF_(short,static)
 		XTAL_LET function(simplex_field_q auto o)
 		noexcept -> decltype(auto)
 		{
